Convert Score and Icon components to TS; move Icons
